backend/pyproject.toml
[project]
name = "photograph-tutor-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.0",
    "livekit-agents>=0.8",
    "livekit-api>=0.7",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
